Revert "e2d61895622975b02b179f20837b682c5c5ad3bb; Add seller manual; add skincare in ideas category."
This reverts commit 170ff25342a1afd9606f91a714bef8be8400b062.

+++ php/actions/comments/delete.php
@@ -5,26 +5,14 @@
  * @package Elgg
  */
 
-// Ensure we're logged in
-if (!elgg_is_logged_in()) {
-	forward();
-}
-
 // Make sure we can get the comment in question
 $annotation_id = (int) get_input('annotation_id');
-if ($comment = elgg_get_annotation_from_id($annotation_id)) {
-
-	$entity = get_entity($comment->entity_guid);
-
-	if ($comment->canEdit()) {
-		$comment->delete();
-		system_message(elgg_echo("generic_comment:deleted"));
-		forward($entity->getURL());
-	}
-
+$comment = elgg_get_annotation_from_id($annotation_id);
+if ($comment && $comment->canEdit()) {
+	$comment->delete();
+	system_message(elgg_echo("generic_comment:deleted"));
 } else {
-	$url = "";
+	register_error(elgg_echo("generic_comment:notdeleted"));
 }
 
-register_error(elgg_echo("generic_comment:notdeleted"));
 forward(REFERER);
